Improved abstraction for overlapping code
- Used an abstract class to place the decompress/compress logic in one place
- Restructure to help maintain
