+++ Call of Duty FastFile Editor/IO/IFastFileHandler.cs
@@ -1,9 +1,4 @@
 ﻿using Call_of_Duty_FastFile_Editor.Models;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 
 namespace Call_of_Duty_FastFile_Editor.IO
 {
